Detection/Detection_Inferentia/neuronocr_deployment/requirements.txt
torch==1.8.1
torchvision==0.9.1
opencv-python
scipy
numpy
Pillow<8.3.0
scikit-image
python-bidi
PyYAML
Flask==2.0.1
waitress==2.0.0




